+++ tests/pom.xml
@@ -32,12 +32,13 @@
 
     <properties>
         <slf4j.version>1.7.30</slf4j.version>
-        <spark.test.version>3.1.0-SNAPSHOT</spark.test.version>
+        <spark.test.version>3.0.0</spark.test.version>
     </properties>
     <profiles>
         <profile>
             <id>spark31</id>
             <properties>
+                <spark.test.version>3.1.0-SNAPSHOT</spark.test.version>
             </properties>
         </profile>
     </profiles>
